+++ CHANGELOG.md
@@ -143,6 +143,7 @@ Modifications can only be done via builders.
 - Allow to add an owner to nodes who is informed about disposal or erasal
 - Allow to add an owner to scopes who is informat about disposal, undisposal and erasal
 - ScBuilderNodeAdder is informed when one of the created nodes is erased
+- Prevent that a builder is applied to scopes created by this builder before
 
 ### Fixed